image/image_tagger-nats/requirements.txt
Pillow
gunicorn
fastapi>=0.61.1
uvicorn>=0.11.8
python-multipart
requests>=2.24.0
open_clip_torch
python-dotenv==1.0.0
transformers >= 4.22.1
sentencepiece >= 0.1.94
datasets >= 1.0.2
sentence-transformers
gdown>=4.7.0
matplotlib
tqdm
opencv-python==4.7.0.72
numpy==1.24.3
pillow>=8.2.0
detectaicore>=0.0.27
orjson
onnxruntime==1.17.0
onnx==1.16.2
psutil
pydantic>=2.0.2
pydantic-settings>=2.0.1
pydantic_core>=2.1.2
nats-py